+++ types/index.d.ts
@@ -1,2 +1,3 @@
-export { default as Modal, bind } from "./Modal.svelte";
-export { default } from "./Modal.svelte";
+export { default as Modal, bind } from './Modal.svelte';
+export { default } from './Modal.svelte';
+export { Context, Open, Close } from './Modal.svelte';